Excellent version to be incorporated into pyFAT-Beta
